商品删除/品牌被商品绑定不能删除

+++ lsj-shop-parent/lsj-shop-service/lsj-shop-service-xxx/src/main/java/com/baidu/shop/service/impl/BrandServiceImpl.java
@@ -5,8 +5,10 @@ import com.baidu.shop.base.Result;
 import com.baidu.shop.dto.BrandDTO;
 import com.baidu.shop.entity.BrandEntity;
 import com.baidu.shop.entity.CategoryBrandEntity;
+import com.baidu.shop.entity.SpuEntity;
 import com.baidu.shop.mapper.BrandMapper;
 import com.baidu.shop.mapper.CategoryBrandMapper;
+import com.baidu.shop.mapper.SpuMapper;
 import com.baidu.shop.service.BrandService;
 import com.baidu.shop.utils.BaiduBeanUtil;
 import com.baidu.shop.utils.ObjectUtil;
@@ -39,6 +41,9 @@ public class BrandServiceImpl extends BaseApiService implements BrandService {
 
     @Resource
     private CategoryBrandMapper categoryBrandMapper;
+
+    @Resource
+    private SpuMapper spuMapper;
 
     @Override
     public Result<PageInfo<BrandEntity>> getBrandInfo(BrandDTO brandDTO) {
@@ -112,7 +117,13 @@ public class BrandServiceImpl extends BaseApiService implements BrandService {
     }
 
     @Override
+    @Transactional
     public Result<JsonObject> deleteBrand(Integer id) {
+        Example example = new Example(SpuEntity.class);
+        example.createCriteria().andEqualTo("brandId",id);
+        List<SpuEntity> list = spuMapper.selectByExample(example);
+        if(list.size() > 0) return this.setResultError("该品牌被商品绑定,不能删除");
+
         brandMapper.deleteByPrimaryKey(id);
         this.deleteCategoryAndBrand(id);
         return this.setResultSuccess();
@@ -126,7 +137,6 @@ public class BrandServiceImpl extends BaseApiService implements BrandService {
 
     //通过brandId删除中间表数据
     private void deleteCategoryAndBrand(Integer id){
-
         Example example = new Example(CategoryBrandEntity.class);
         example.createCriteria().andEqualTo("brandId",id);
         categoryBrandMapper.deleteByExample(example);
